Fix sanitization routines for PHP 7.1
Lines 88 and 188 set `$output` to an empty string, and subsequent lines in these `cyberchimps_sanitize_multicheck` and `cyberchimps_sanitize_section_order` add items to these lists as if they were arrays. In PHP 7.1, this triggers a fatal error:

```
Fatal error: Uncaught Error: [] operator not supported for strings
```

This commit fixes the issue.